SHIELD/requirements.txt
paho-mqtt==1.6.1
keyboard==0.13.5
networkx==3.3
pandas==2.2.2
matplotlib==3.8.4
nvdlib==0.7.6